% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{abundances}
\alias{abundances}
\title{Get the abundance matrix}
\usage{
abundances(object, ...)
}
\arguments{
\item{object}{the object to get from}

\item{...}{additional arguments used by the concrete implementation}
}
\description{
Get the abundance matrix
}
\seealso{
\link{accessor_methods} for the implementation for a 'proDAFit' object
}
