% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{accessor_methods}
\alias{accessor_methods}
\alias{abundances,proDAFit-method}
\alias{design,proDAFit-method}
\alias{hyper_parameters,proDAFit-method}
\alias{feature_parameters,proDAFit-method}
\alias{coefficients,proDAFit-method}
\alias{coefficient_variance_matrices,proDAFit-method}
\alias{reference_level,proDAFit-method}
\alias{convergence,proDAFit-method}
\title{Get different features and elements of the 'proDAFit' object}
\usage{
\S4method{abundances}{proDAFit}(object)

\S4method{design}{proDAFit}(object, formula = FALSE)

\S4method{hyper_parameters}{proDAFit}(object)

\S4method{feature_parameters}{proDAFit}(object)

\S4method{coefficients}{proDAFit}(object)

\S4method{coefficient_variance_matrices}{proDAFit}(object)

\S4method{reference_level}{proDAFit}(object)

\S4method{convergence}{proDAFit}(object)
}
\arguments{
\item{object}{the 'proDAFit' object}

\item{formula}{specific argument for the \code{design} function
to get the formula that was used to create the linear model.
If no formula was used \code{NULL} is returned.}
}
\description{
The functions listed below can all be accessed using the
fluent dollar notation (ie. \code{fit$abundances[1:3,1:3]}) without
any additional parentheses.
}
