% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{coefficient_variance_matrices}
\alias{coefficient_variance_matrices}
\title{Get the coefficients}
\usage{
coefficient_variance_matrices(object, ...)
}
\arguments{
\item{object}{the object to get from}

\item{...}{additional arguments used by the concrete implementation}
}
\description{
Get the coefficients
}
\seealso{
\link{accessor_methods} for the implementation for a 'proDAFit' object
}
