% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{coefficients,proDAFit-method}
\alias{coefficients,proDAFit-method}
\title{Get the coefficients as a matrix}
\usage{
\S4method{coefficients}{proDAFit}(object)
}
\description{
Get the coefficients as a matrix
}
