% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{convergence,proDAFit-method}
\alias{convergence,proDAFit-method}
\title{Get information on the convergence}
\usage{
\S4method{convergence}{proDAFit}(object)
}
\description{
Get information on the convergence
}
