% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{convergence}
\alias{convergence}
\title{Get the convergence information}
\usage{
convergence(object, ...)
}
\arguments{
\item{object}{the object to get from}

\item{...}{additional arguments used by the concrete implementation}
}
\value{
a list with information on the convergence
}
\description{
Get the convergence information
}
\examples{
  syn_data <- generate_synthetic_data(n_proteins = 10)
  fit <- proDA(syn_data$Y, design = syn_data$groups)
  convergence(fit)

}
\seealso{
\link{accessor_methods} for the implementation for a 'proDAFit' object
}
