% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{design,proDAFit-method}
\alias{design,proDAFit-method}
\title{Access the design matrix or formula used to fit the parameters}
\usage{
\S4method{design}{proDAFit}(object, formula = FALSE)
}
\description{
Access the design matrix or formula used to fit the parameters
}
