% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\docType{methods}
\name{dist_approx,proDAFit-method}
\alias{dist_approx,proDAFit-method}
\title{Distance method for 'proDAFit' object}
\usage{
\S4method{dist_approx}{proDAFit}(object, by_sample = TRUE)
}
\description{
Distance method for 'proDAFit' object
}
