% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{feature_parameters,proDAFit-method}
\alias{feature_parameters,proDAFit-method}
\title{Get a feature parameters as a data.frame}
\usage{
\S4method{feature_parameters}{proDAFit}(object)
}
\description{
Get a feature parameters as a data.frame
}
