% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{feature_parameters}
\alias{feature_parameters}
\title{Get the feature parameters}
\usage{
feature_parameters(object, ...)
}
\arguments{
\item{object}{the object to get from}

\item{...}{additional arguments used by the concrete implementation}
}
\value{
a data.frame with information on each fit
}
\description{
Get the feature parameters
}
\examples{
  syn_data <- generate_synthetic_data(n_proteins = 10)
  fit <- proDA(syn_data$Y, design = syn_data$groups)
  feature_parameters(fit)

}
\seealso{
\link{accessor_methods} for the implementation for a 'proDAFit' object
}
