% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_parameters.R
\name{fit_parameters}
\alias{fit_parameters}
\title{Main proDA function to determine the hyper and protein parameters}
\usage{
fit_parameters(X, design = `???`)
}
\description{
Main proDA function to determine the hyper and protein parameters
}
