% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%zero_dom_mat_mult\%}
\alias{\%zero_dom_mat_mult\%}
\title{Helper function that makes sure that NA * 0 = 0 in matrix multiply}
\usage{
X \%zero_dom_mat_mult\% Y
}
\description{
Helper function that makes sure that NA * 0 = 0 in matrix multiply
}
\keyword{internal}
