% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{hyper_parameters,proDAFit-method}
\alias{hyper_parameters,proDAFit-method}
\title{Get a list of the hyperparameters}
\usage{
\S4method{hyper_parameters}{proDAFit}(object)
}
\description{
Get a list of the hyperparameters
}
