% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{hyper_parameters}
\alias{hyper_parameters}
\title{Get the hyper parameters}
\usage{
hyper_parameters(object, ...)
}
\arguments{
\item{object}{the object to get from}

\item{...}{additional arguments used by the concrete implementation}
}
\description{
Get the hyper parameters
}
\seealso{
\link{accessor_methods} for the implementation for a 'proDAFit' object
}
