% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{invprobit_fast}
\alias{invprobit_fast}
\title{Same thing as invprobit, but without the parameter validation}
\usage{
invprobit_fast(x, rho, zeta, log = FALSE, oneminus = FALSE)
}
\value{
a numeric vector of \code{length(x)}
}
\description{
Same thing as invprobit, but without the parameter validation
}
\keyword{internal}
