% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_lm.R
\name{pd_lm}
\alias{pd_lm}
\title{Fit linear probabilistic dropout model}
\usage{
pd_lm(y, X, rho, zeta, mu0, sigma20, df0, tau20, location_prior_df = 3,
  method = c("analytic_grad", "numeric_grad"))
}
\description{
Fit linear probabilistic dropout model
}
