% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,proDAFit-method}
\alias{predict,proDAFit-method}
\title{Predict the parameters or values of additional proteins}
\usage{
\S4method{predict}{proDAFit}(object, newdata, type = c("response",
  "feature_parameters"), ...)
}
\description{
Predict the parameters or values of additional proteins
}
