% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proDA.R
\docType{package}
\name{proDA}
\alias{proDA}
\alias{proDA-package}
\title{proDA: Identify differentially abundant proteins in label-free mass spectrometry}
\description{
proDA: Identify differentially abundant proteins in label-free mass spectrometry
}
