% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{reference_level,proDAFit-method}
\alias{reference_level,proDAFit-method}
\title{Get a the reference class}
\usage{
\S4method{reference_level}{proDAFit}(object)
}
\description{
Get a the reference class
}
