% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{reference_level}
\alias{reference_level}
\title{Get the reference level}
\usage{
reference_level(object, ...)
}
\arguments{
\item{object}{the object to get from}

\item{...}{additional arguments used by the concrete implementation}
}
\value{
a string
}
\description{
Get the reference level
}
\examples{
  syn_data <- generate_synthetic_data(n_proteins = 10)
  fit <- proDA(syn_data$Y, design = syn_data$groups)
  reference_level(fit)

}
\seealso{
\link{accessor_methods} for the implementation for a 'proDAFit' object
}
