% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{abundances}
\alias{abundances}
\title{Get the abundance matrix}
\usage{
abundances(object, ...)
}
\arguments{
\item{object}{the object to get from}

\item{...}{additional arguments used by the concrete implementation}
}
\value{
the original matrix that was fitted
}
\description{
Get the abundance matrix
}
\examples{
  syn_data <- generate_synthetic_data(n_proteins = 10)
  fit <- proDA(syn_data$Y, design = syn_data$groups)
  abundances(fit)

}
\seealso{
\link{accessor_methods} for the implementation for a 'proDAFit' object
}
