% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{coefficients}
\alias{coefficients}
\title{Get the coefficients}
\usage{
coefficients(object, ...)
}
\arguments{
\item{object}{the object to get from}

\item{...}{additional arguments used by the concrete implementation}
}
\value{
a numeric matrix of size `nrow(fit) * p`
}
\description{
Get the coefficients
}
\examples{
  syn_data <- generate_synthetic_data(n_proteins = 10)
  fit <- proDA(syn_data$Y, design = syn_data$groups)
  coefficients(fit)

}
\seealso{
\link{accessor_methods} for the implementation for a 'proDAFit' object
}
