% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{convergence}
\alias{convergence}
\title{Get the convergence information}
\usage{
convergence(object, ...)
}
\arguments{
\item{object}{the object to get from}

\item{...}{additional arguments used by the concrete implementation}
}
\description{
Get the convergence information
}
\seealso{
\link{accessor_methods} for the implementation for a 'proDAFit' object
}
