% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{dist_approx}
\alias{dist_approx}
\title{Calculate an approximate distance for 'object'}
\usage{
dist_approx(object, ...)
}
\arguments{
\item{object}{the object for which the distance is approximated}

\item{...}{additional arguments used by the concrete implementation}
}
\description{
Calculate an approximate distance for 'object'
}
\seealso{
\code{\link[stats]{dist}} for the base R function and
  \code{\link[dist_approx,proDAFit-method]{dist_approx()}} concrete implementation
  for 'proDAFit' objects
}
