% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{feature_parameters}
\alias{feature_parameters}
\title{Get the feature parameters}
\usage{
feature_parameters(object, ...)
}
\arguments{
\item{object}{the object to get from}

\item{...}{additional arguments used by the concrete implementation}
}
\description{
Get the feature parameters
}
\seealso{
\link{accessor_methods} for the implementation for a 'proDAFit' object
}
