% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_parameters.R
\name{fit_parameters}
\alias{fit_parameters}
\title{Main proDA function to determine the hyper and protein parameters}
\usage{
fit_parameters(data, design = ~1, col_data = NULL,
  reference_level = NULL, data_is_log_transformed = TRUE,
  max_iter = 20)
}
\description{
Main proDA function to determine the hyper and protein parameters
}
