% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\docType{methods}
\name{median_normalization}
\alias{median_normalization}
\alias{median_normalization,SummarizedExperiment-method}
\title{Column wise median normalization of the data matrix}
\usage{
median_normalization(X)

\S4method{median_normalization}{SummarizedExperiment}(X)
}
\arguments{
\item{X}{a matrix of proteins and samples}
}
\value{
the normalized matrix
}
\description{
The method calculates for each sample the median change (i.e. the difference
between the observed value and the row average) and substracts it from
each row. Missing values are ignored in the procedure. The method is based
on the assumption that a majority of the rows did not change.
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: S4 method of \code{median_normalization} for
\code{SummarizedExperiment}
}}

