% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_lm.R
\name{pd_lm}
\alias{pd_lm}
\title{Fit linear probabilistic dropout model}
\usage{
pd_lm(formula, data, subset, weights, method = c("analytic_grad",
  "numeric_grad"))
}
\description{
Fit linear probabilistic dropout model
}
