% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proDA.R
\name{proDA}
\alias{proDA}
\title{Main proDA function to determine the hyper and protein parameters}
\usage{
proDA(data, design = ~1, col_data = NULL, reference_level = NULL,
  data_is_log_transformed = TRUE, location_prior_df = 3,
  moderate_location = TRUE, moderate_variance = TRUE, max_iter = 20,
  epsilon = 0.001, verbose = FALSE, ...)
}
\description{
Main proDA function to determine the hyper and protein parameters
}
