% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_diff.R
\name{test_diff}
\alias{test_diff}
\title{Main proDA function to identify significant parameters}
\usage{
test_diff(parameters, contrast = `???`)
}
\description{
Main proDA function to identify significant parameters
}
